requirements.txt
flask==2.3.2
mysql-connector-python==8.1.0
flask_cors==4.0.0
Werkzeug==2.3.3
requests==2.31.0